Move commander commands to sub-files.
