--- packages/cli/src/database/index.ts
@@ -1,3 +1,0 @@
-export { createGroup } from './create-group';
-export { createPermission } from './create-permission';
-export { createUser } from './create-user';
