--- packages/cli/src/generate/index.ts
@@ -1,1 +1,0 @@
-export * from './generators';
